requirements.txt
fastapi
uvicorn[standard]
pymongo
sentence-transformers
python-dotenv
